+++ src/cache/OntologyCache.ts
@@ -1,6 +1,6 @@
 import { AsyncSubject, forkJoin, Observable, of } from "rxjs";
 import { map, mergeMap } from "rxjs/operators";
-import { KnoraApiConfig } from "..";
+import { KnoraApiConfig } from "../knora-api-config";
 import { V2Endpoint } from "../api/v2/v2-endpoint";
 import { ClassDefinition, IHasProperty } from "../models/v2/ontologies/class-definition";
 import { OntologyConversionUtil } from "../models/v2/ontologies/OntologyConversionUtil";
